+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.0
 
 New features:
-  * Add env parameter in request header.
+  * Add extra fields in request header.
   * Setup via provided configuration files and values in the parameters.yml
 
 ## 1.0.0